feat: ゲーム進行結果の型とスキーマを新規作成し、sendCommandメソッドの戻り値を更新
